+++ axum-extra/src/lib.rs
@@ -44,6 +44,5 @@
 #![cfg_attr(test, allow(clippy::float_cmp))]
 
 pub mod extract;
-pub mod middleware;
 pub mod response;
 pub mod routing;