+++ axum-extra/src/middleware/mod.rs
@@ -1,5 +1,0 @@
-//! Additional types for creating middleware.
-
-pub mod middleware_fn;
-
-pub use self::middleware_fn::{from_fn, Next};